+++ packages/providers/test/provider.test.ts
@@ -11,7 +11,7 @@ import * as fuelTsVersionsMod from '@fuel-ts/versions';
 import { getBytesCopy, hexlify } from 'ethers';
 import type { BytesLike } from 'ethers';
 
-import type { FetchRequestOptions } from '../src/provider';
+import type { TransactionCost, FetchRequestOptions } from '../src/provider';
 import Provider from '../src/provider';
 import type {
   CoinTransactionRequestInput,
@@ -860,18 +860,21 @@ describe('Provider', () => {
     const provider = await Provider.create(FUEL_NETWORK_URL);
     const gasLimit = 1;
     const gasUsed = bn(1000);
-    const transactionParams = {
-      minGasPrice: bn(1),
-      gasPrice: bn(1),
+    const transactionCost: TransactionCost = {
       gasUsed,
-      fee: bn(1),
+      gasPrice: bn(1),
+      minGasPrice: bn(1),
+      maxFee: bn(2),
+      minFee: bn(1),
+      receipts: [],
+      requiredQuantities: [],
     };
 
     const estimateTxSpy = jest.spyOn(provider, 'estimateTxDependencies').mockImplementation();
 
     const txCostSpy = jest
       .spyOn(provider, 'getTransactionCost')
-      .mockReturnValue(Promise.resolve(transactionParams));
+      .mockReturnValue(Promise.resolve(transactionCost));
 
     await expectToThrowFuelError(
       () => provider.sendTransaction(new ScriptTransactionRequest({ gasPrice: 1, gasLimit })),
@@ -889,18 +892,21 @@ describe('Provider', () => {
     const provider = await Provider.create(FUEL_NETWORK_URL);
     const gasPrice = 1;
     const minGasPrice = bn(1000);
-    const transactionParams = {
+    const transactionCost: TransactionCost = {
       minGasPrice,
       gasPrice: bn(1),
       gasUsed: bn(1),
-      fee: bn(1),
+      maxFee: bn(2),
+      minFee: bn(1),
+      receipts: [],
+      requiredQuantities: [],
     };
 
     const estimateTxSpy = jest.spyOn(provider, 'estimateTxDependencies').mockImplementation();
 
     const txCostSpy = jest
       .spyOn(provider, 'getTransactionCost')
-      .mockReturnValue(Promise.resolve(transactionParams));
+      .mockReturnValue(Promise.resolve(transactionCost));
 
     await expectToThrowFuelError(
       () => provider.sendTransaction(new ScriptTransactionRequest({ gasPrice, gasLimit: 1000 })),