+++ packages/script/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+## 0.67.0
+
+### Minor Changes
+
+- 🐞 Fixing transaction funding, by [@Torres-ssf](https://github.com/Torres-ssf) (See [#1372](https://github.com/FuelLabs/fuels-ts/pull/1372))
+
 ## 0.66.1
 
 ### Patch Changes